+++ pkg/execution/query/iterator_test.go
@@ -464,7 +464,7 @@ func mustCreateTupleDesc() *tuple.TupleDescription {
 	return td
 }
 
-func createTestTuple(td *tuple.TupleDescription, id int32, name string) *tuple.Tuple {
+func createTestTuple(td *tuple.TupleDescription, id int64, name string) *tuple.Tuple {
 	t := tuple.NewTuple(td)
 	intField := types.NewIntField(id)
 	stringField := types.NewStringField(name, 128)
