+++ pkg/execution/query/seqscan_test.go
@@ -258,7 +258,7 @@ func mustCreateSeqScanTupleDesc() *tuple.TupleDescription {
 	return td
 }
 
-func createSeqScanTestTuple(td *tuple.TupleDescription, id int32, name string) *tuple.Tuple {
+func createSeqScanTestTuple(td *tuple.TupleDescription, id int64, name string) *tuple.Tuple {
 	t := tuple.NewTuple(td)
 	intField := types.NewIntField(id)
 	stringField := types.NewStringField(name, 128)
